Manually 'play()' the stream element as soon as possible.
This commit changea the 'autoplay' feature of the streaming transcribed element,
to deal with Firefox and the Speech SDK. The Speech SDK, by default, uses a
'canplaythrough' event to trigger transcription. Current Firefox interprets
'canplaythrough' as the entire file needing to be downloaded, which delays the
emission of this event.